requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
tensorflow-cpu
gensim

# tests/linter
# black
# coverage
# flake8
# pytest
# yapf

# API
gcsfs
google-cloud-storage
s3fs

# utilities
pillow
streamlit
six>=1.14
